Allow passing a platform argument to generate graph for CLJS
This required an upgrade to tools.namespace too. The new version of
tools.namespace requires tools.reader 0.10.0, but when I was testing
this there was another version of tools.reader coming from somewhere
which was conflicting. I've opened a ticket against Leiningen to upgrade
their clj-http dependency as I suspect it could be this one.

I've vendored read-file-ns-decl so that it prints out exceptions that
get thrown. I'm not 100% sure if this is the right option, but I don't
like how it swallows all exceptions.

Fixes #1